refactor: custom matchers trasladados a la clase MatcherUtils
