+++ app/src/androidTest/java/ar/edu/unq/pdes/myprivateblog/PostEditTest.kt
@@ -1,15 +1,9 @@
 package ar.edu.unq.pdes.myprivateblog
 
 import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
-import android.os.IBinder
-import android.view.View
-import android.view.WindowManager
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.Root
 import androidx.test.espresso.action.ViewActions.*
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.BoundedMatcher
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.espresso.web.assertion.WebViewAssertions.webMatches
 import androidx.test.espresso.web.sugar.Web.onWebView
@@ -19,6 +13,8 @@ import androidx.test.espresso.web.webdriver.Locator
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
+import ar.edu.unq.pdes.myprivateblog.MatcherUtils.Companion.withBackgroundColor
+import ar.edu.unq.pdes.myprivateblog.MatcherUtils.Companion.withTintColor
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.containsString
@@ -136,53 +132,11 @@ class PostEditTest {
 
         onView(withId(R.id.title)).check(matches(hasErrorText("Debe tener algún título")))
 
-        onView(withText("Error al guardar el post")).inRoot(ToastMatcher())
+        onView(withText("Error al guardar el post")).inRoot(MatcherUtils.ToastMatcher())
             .check(matches(isDisplayed()))
     }
 
-    private fun withBackgroundColor(colorPicked: Int): Matcher<View?> {
-        return object : BoundedMatcher<View?, View>(View::class.java) {
-            public override fun matchesSafely(view: View): Boolean {
-                val color = view.background.current as ColorDrawable
-                return color.color == colorPicked
-            }
-            override fun describeTo(description: Description) {
-                description.appendText("Checking the matcher on received view: ")
-                description.appendText("with expectedColor = $colorPicked")
-            }
-        }
-    }
 
 
-    private fun withTintColor(expectedColor: Int): Matcher<View?> {
-        return object : BoundedMatcher<View?, View>(View::class.java) {
-            public override fun matchesSafely(view: View): Boolean {
-                return view.backgroundTintList?.defaultColor == expectedColor
-            }
-            override fun describeTo(description: Description) {
-                description.appendText("Checking the matcher on received view: ")
-                description.appendText("with expectedColor = $expectedColor")
-            }
-        }
-    }
 
-    class ToastMatcher : TypeSafeMatcher<Root?>() {
-
-        override fun describeTo(description: Description?) {
-            description?.appendText("is toast")
-        }
-
-        override fun matchesSafely(item: Root?): Boolean {
-            val type: Int? = item?.windowLayoutParams?.get()?.type
-            if (type == WindowManager.LayoutParams.TYPE_TOAST) {
-                val windowToken: IBinder = item.decorView.windowToken
-                val appToken: IBinder = item.decorView.applicationWindowToken
-                if (windowToken === appToken) { // means this window isn't contained by any other windows.
-                    return true
-                }
-            }
-            return false
-        }
-
-    }
 }
