+++ app/src/androidTest/java/ar/edu/unq/pdes/myprivateblog/PostsListingTest.kt
@@ -1,18 +1,15 @@
 package ar.edu.unq.pdes.myprivateblog
 
 import android.graphics.Color
-import android.view.View
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.typeText
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.BoundedMatcher
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import org.hamcrest.Description
-import org.hamcrest.Matcher
+import ar.edu.unq.pdes.myprivateblog.MatcherUtils.Companion.withTintColor
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -75,19 +72,7 @@ class PostsListingTest {
 
 }
 
-fun withTintColor(expectedColor: Int): Matcher<View?>? {
-    return object : BoundedMatcher<View?, View>(View::class.java) {
 
-        override fun describeTo(description: Description) {
-            description.appendText("Checking the matcher on received view: ")
-            description.appendText("with expectedStatus=$expectedColor")
-        }
-
-        override fun matchesSafely(view: View): Boolean {
-            return view.backgroundTintList?.defaultColor == expectedColor
-        }
-    }
-}
 //
 //fun withWebViewTextMatcher(expectedText: String): Matcher<View?>? {
 //    return object : BoundedMatcher<View?, WebView>(WebView::class.java) {
